player class generates events when one of its core variables are modified. buytower scren now handles the data from a state change to it directly using fully persistent event delegates. playing screen has an incorporated sell tower button if you click on a tower that is already placed

+++ include/GameLogic.hpp
@@ -38,9 +38,9 @@ private:
   //y dimension of the render window
   int windowY;
   //x dimension in pixels of a grid space
-  int gridX;
+  float gridX;
   //y dimension in pixels of a grid space
-  int gridY;
+  float gridY;
   //Store the textLoader to make requests for strings and constants
   shared_ptr<TextLoader> textLoader;
   //Store the game's instance of the BoardManager class that the game logic
@@ -122,8 +122,8 @@ public:
   bool isExit(int row, int col);
   bool isPath(int row, int col);
 
-  const int getTileXSize();
-  const int getTileYSize();
+  const float getTileXSize();
+  const float getTileYSize();
   const int getRows();
   const int getCols();
   const int getWindowY();
